Fixed a bug with the 'devblocks_prettybytes' template modifier where it would overflow a signed 4-byte integer on 32-bit filesystems.  Swapped intval() to floatval().  64-bit systems were not affected.

+++ Devblocks.class.php
@@ -3450,8 +3450,8 @@ class _DevblocksTemplateManager {
 		if(!is_numeric($string))
 			return '';
 			
-		$bytes = intval($string);
-		$precision = intval($precision);
+		$bytes = floatval($string);
+		$precision = floatval($precision);
 		$out = '';
 		
 		if($bytes >= 1000000000) {
